requirements.txt
telebot
openai
Flask
PyYAML
requests
